validation/advanced:  start to add EB FT test

+++ validation/advanced-tests/run-eb-tests.sh
@@ -1,6 +1,13 @@
 #!/bin/sh -f
 
+webDir=http://clasweb.jlab.org/clas12offline/distribution/coatjava/validation_files/eb
+webVersion=v0
+webDir=$webDir/$webVersion
+
 # coatjava must already be built at ../../coatjava/
+
+# whether to use CLARA (0=no)
+useClara=0
 
 # if first argument is -t, only run the test, 
 # don't redownload dependencies, don't run reconstruction.
@@ -19,24 +26,24 @@ case $webFileStub in
         ;;
     electronpion)
         ;;
+    forwardtagger)
+        ;;
     *)
       echo Invalid input evio file:  $webFileStub
       exit 1
 esac
 
 # set up environment
-CLARA_HOME=$PWD/clara_installation/ ; export CLARA_HOME
-
-# Do NOT use whatever clas12 libraries come with CLARA for testing:
-#COAT=$CLARA_HOME/plugins/clas12/
-
-# Use the ones in this clas12 build for testing:
-COAT=../../coatjava
+if [ $useClara -eq 0 ]
+then
+    COAT=../../coatjava
+else
+    CLARA_HOME=$PWD/clara_installation/
+    COAT=$CLARA_HOME/plugins/clas12/
+    export CLARA_HOME
+fi
 
 classPath="$COAT/lib/services/*:$COAT/lib/clas/*:$COAT/lib/utils/*:../lib/*:src/"
-
-webDir=http://clasweb.jlab.org/clas12offline/distribution/coatjava/validation_files/eb/v0/
-    
 
 # compile test codes before anything else:
 javac -cp $classPath src/eb/EBTwoTrackTest.java
@@ -45,20 +52,24 @@ if [ $? != 0 ] ; then echo "EBTwoTrackTest compilation failure" ; exit 1 ; fi
 # download and setup dependencies, run reconstruction:
 if [ $runTestOnly -eq 0 ]
 then
-    # tar the local coatjava build so it can be installed with clara
-    cd ../..
-    tar -zcvf coatjava-local.tar.gz coatjava
-    mv coatjava-local.tar.gz validation/advanced-tests/
-    cd -
 
-    # install clara
-    if ! [ -d clara_installation ]
+    if ! [ $useClara -eq 0 ]
     then
-        wget --no-check-certificate https://claraweb.jlab.org/clara/_downloads/install-claracre-clas.sh
-        chmod +x install-claracre-clas.sh
-        ./install-claracre-clas.sh -l local
-        if [ $? != 0 ] ; then echo "clara installation error" ; exit 1 ; fi
-        rm install-claracre-clas.sh
+        # tar the local coatjava build so it can be installed with clara
+        cd ../..
+        tar -zcvf coatjava-local.tar.gz coatjava
+        mv coatjava-local.tar.gz validation/advanced-tests/
+        cd -
+
+        # install clara
+        if ! [ -d clara_installation ]
+        then
+            wget --no-check-certificate https://claraweb.jlab.org/clara/_downloads/install-claracre-clas.sh
+            chmod +x install-claracre-clas.sh
+            ./install-claracre-clas.sh -l local
+            if [ $? != 0 ] ; then echo "clara installation error" ; exit 1 ; fi
+            rm install-claracre-clas.sh
+        fi
     fi
 
     # download test files
@@ -72,27 +83,29 @@ then
     rm -f ${webFileStub}.hipo
     rm -f out_${webFileStub}.hipo
 
-    # convert to hipo
+    # convert to hipo:
     $COAT/bin/evio2hipo -o ${webFileStub}.hipo ${webFileStub}.evio
 
-    # run reconstruction without clara
-    ../../coatjava/bin/notsouseful-util -i ${webFileStub}.hipo -o out_${webFileStub}.hipo -c 2
-
-    # run reconstruction with clara
-    #echo "set inputDir $PWD/" > cook.clara
-    #echo "set outputDir $PWD/" >> cook.clara
-    #echo "set threads 7" >> cook.clara
-    #echo "set javaMemory 2" >> cook.clara
-    #echo "set session s_cook" >> cook.clara
-    #echo "set description d_cook" >> cook.clara
-    #ls ${webFileStub}.hipo > files.list
-    #echo "set fileList $PWD/files.list" >> cook.clara
-    #echo "run local" >> cook.clara
-    #echo "exit" >> cook.clara
-    #$CLARA_HOME/bin/clara-shell cook.clara
+    # run reconstruction:
+    if [ $useClara -eq 0 ]
+    then
+        ../../coatjava/bin/notsouseful-util -i ${webFileStub}.hipo -o out_${webFileStub}.hipo -c 2
+    else
+        echo "set inputDir $PWD/" > cook.clara
+        echo "set outputDir $PWD/" >> cook.clara
+        echo "set threads 7" >> cook.clara
+        echo "set javaMemory 2" >> cook.clara
+        echo "set session s_cook" >> cook.clara
+        echo "set description d_cook" >> cook.clara
+        ls ${webFileStub}.hipo > files.list
+        echo "set fileList $PWD/files.list" >> cook.clara
+        echo "run local" >> cook.clara
+        echo "exit" >> cook.clara
+        $CLARA_HOME/bin/clara-shell cook.clara
+    fi
 fi
 
-# run KppTracking junit tests
+# run Event Builder tests:
 java -DCLAS12DIR="$COAT" -Xmx1536m -Xms1024m -cp $classPath -DINPUTFILE=out_${webFileStub}.hipo org.junit.runner.JUnitCore eb.EBTwoTrackTest
 if [ $? != 0 ] ; then echo "EBTwoTrackTest unit test failure" ; exit 1 ; else echo "EBTwoTrackTest passed unit tests" ; fi
